fix broker<cloud sensor

+++ leshan-server-demo/src/main/java/org/eclipse/leshan/server/demo/AvailableLightDiscovery.java
@@ -18,11 +18,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class AvailableLightDiscovery {
-	
+
 	private static final String FORMAT_PARAM = "format";
 
     private static final Logger LOG = LoggerFactory.getLogger(AvailableLightDiscovery.class);
-    
+
     public static final String LIGHT_STATE = "/10250/0/2";
     public static final String LIGHT_USER_TYPE = "/10250/0/3";
     public static final String LIGHT_USER_ID = "/10250/0/4";
@@ -32,30 +32,30 @@ public class AvailableLightDiscovery {
     public static final String SENSOR_USER_ID = "/10350/0/3";
     public static final String SENSOR_LOCATION_X = "/10250/0/5";
     public static final String SENSOR_LOCATION_Y = "/10250/0/6";
-    
+
     public final static String obervables = "/10250/0/2;/10250/0/3;/10250/0/4;/10350/0/2;/10350/0/3";
     public final static String locationInfo = "/10250/0/8;/10250/0/9;;/10350/0/5;/10350/0/6";
-    
-    
+
+
     private static final long TIMEOUT = 5000; // ms
-    
+
     public static HashMap<String, HashMap<String, Object>> devices;
-    
+
     private static BrokerServlet broker;
-    
+
 	private final LwM2mServer server;
-	
+
 	 RegistrationListener registrationListener = new RegistrationListener() {
-			
+
 			@Override
 			public void updated(RegistrationUpdate update, Registration updatedRegistration) {}
-			
+
 			@Override
 			public void unregistered(Registration registration) {
 				// TODO remove device from lists
-				
+
 			}
-			
+
 			@Override
 			public void registered(Registration registration) {
 				// observe the state of the devices for the Available Light Discovery
@@ -107,9 +107,9 @@ public class AvailableLightDiscovery {
 				}
 			}
 		};
-		
+
 		ObservationListener observationListener = new ObservationListener() {
-			
+
 			@Override
 			public void newValue(Observation observation, ObserveResponse response) {
 				Registration registration = server.getRegistrationService().getById(observation.getRegistrationId());
@@ -122,10 +122,10 @@ public class AvailableLightDiscovery {
 	            	devices.put(registration.getEndpoint(), resources);
 	            }
 			}
-			
+
 			@Override
 			public void newObservation(Observation observation) {}
-			
+
 			@Override
 			public void cancelled(Observation observation) {
 				//registration is always null so this option does not work !!
@@ -136,31 +136,31 @@ public class AvailableLightDiscovery {
 //						try {
 //							ObserveRequest obsReq = new ObserveRequest(ContentFormat.JSON, resource);
 //							ObserveResponse cResponse = server.send(registration, obsReq, TIMEOUT);
-//						} catch (InterruptedException e) {						
+//						} catch (InterruptedException e) {
 //							e.printStackTrace();
 //						}
-//						
+//
 //					}
 //				}
 			}
 		};
-	
+
 	public AvailableLightDiscovery(LwM2mServer server) {
 		this.server = server;
         devices = new HashMap<>();
-        
+
 		this.server.getRegistrationService().addListener(this.registrationListener);
 		this.server.getObservationService().addListener(this.observationListener);
 	}
-	
+
 	public void setBroker(BrokerServlet broker) {
 		this.broker = broker;
 	}
-	
+
 	public static  Boolean userAtDesk(String userID) {
 		if(userID.equals("Office-Admin-0")) return true;
 		User user = broker.usersMap.get(userID);
-		
+
 		//check if required resources are known
 		if(!devices.containsKey(user.getSensor())) return false;
 		if(!devices.get(user.getSensor()).containsKey(SENSOR_STATE)) return false;
@@ -168,38 +168,38 @@ public class AvailableLightDiscovery {
 		return devices.get(user.getSensor()).get(SENSOR_STATE).equals("OCCUPIED");
 //		return devices.get(user.getSensor()).get(SENSOR_STATE).equals("USED");
 	}
-	
+
 	public Boolean isUserCloserToLight(String userID, String userID2, String endpoint) {
 		User user = broker.usersMap.get(userID);
 		User user2 = broker.usersMap.get(userID2);
-		
+
 		//check if required resources are known
 		if(!devices.containsKey(endpoint)) return false;
 		if(!devices.get(endpoint).containsKey(LIGHT_LOCATION_X)) return false;
 		if(!devices.get(endpoint).containsKey(LIGHT_LOCATION_Y)) return false;
-		
-		if(user.getDistanceFromLocation((int) devices.get(endpoint).get(LIGHT_LOCATION_X), (int) devices.get(endpoint).get(LIGHT_LOCATION_Y)) < 
-			user2.getDistanceFromLocation((int) devices.get(endpoint).get(LIGHT_LOCATION_X), (int) devices.get(endpoint).get(LIGHT_LOCATION_Y)))
+
+		if(user.getDistanceFromLocation((double) devices.get(endpoint).get(LIGHT_LOCATION_X), (double) devices.get(endpoint).get(LIGHT_LOCATION_Y)) <
+			user2.getDistanceFromLocation((double) devices.get(endpoint).get(LIGHT_LOCATION_X), (double) devices.get(endpoint).get(LIGHT_LOCATION_Y)))
 			return true;
 		return false;
 	}
-	
+
 	public Boolean isAvailabletoUser(String endpoint, String userID) {
 		User user = broker.usersMap.get(userID);
 		// User is USER1 of this light
 		if (user.getLightUSER1().equals(endpoint)) return true;
-		
+
 		//check if required resources are known
 		if(!devices.containsKey(endpoint)) return false;
 		if(!devices.get(endpoint).containsKey(LIGHT_STATE)) return false;
 		if(!devices.get(endpoint).containsKey(LIGHT_USER_TYPE)) return false;
-		
-		
+
+
 		// User is USER2 && USER1 is not at his desk
-		if (user.getLightUSER2().equals(endpoint) && (devices.get(endpoint).get(LIGHT_STATE).equals("FREE") ||  
+		if (user.getLightUSER2().equals(endpoint) && (devices.get(endpoint).get(LIGHT_STATE).equals("FREE") ||
 			(devices.get(endpoint).get(LIGHT_STATE).equals("USED") && !devices.get(endpoint).get(LIGHT_USER_TYPE).equals("USER1"))))
 			return true;
-		
+
 		//check if required resources are known
 		if(!devices.get(endpoint).containsKey(LIGHT_USER_ID)) return false;
 		// User is USER3
@@ -207,8 +207,8 @@ public class AvailableLightDiscovery {
 		  (devices.get(endpoint).get(LIGHT_STATE).equals("USED") && devices.get(endpoint).get(LIGHT_USER_TYPE).equals("USER3") &&
 		    isUserCloserToLight(userID, (String) devices.get(endpoint).get(LIGHT_USER_ID), endpoint)))
 			return true;
-		
-		
+
+
 		return false;
 	}
 }
